Stripe: Support verify_credentials
We'll gradually be adding support for this operation to all gateways.
So far, we've implemented it in BlueSnap and now Stripe.

Previously, customers would need to run a real transaction to determine
if the api credentials for a gateway were legitimate.  Now we can ask a
gateway directly without needing to give a credit card.