Clearhaus: Fix refund of captures
Refunds of purchases were working since we were using the
`:use_first_response` of the MultiResponse in `purchase` which gives us
the id of the original auth.  Refunds of captures were broken though
since we were sending the ID of the capture rather than the ID of the
auth which is what Clearhaus wants.  A 404 is returned if you attempt to
use the id of the capture.

Successful Full Test Suite Run at 2016-07-07 13:56:55