SagePay: Fix truncation
SagePay has size limits on most of their fields.  The problem is, they
impose those limits on the url encoded string.  We were thinking that we
could truncate the value, url encode it, and send it to the server
such that they could impose the size limits after the value was url
decoded.

But alas, they impose the limits on the url encoded value.