Sage: Default billing state when outside US
Sage requires an address state be sent even if transacting outside of
the US/North America and the country does not have states.
Unfortunately Sage support is unwilling to say what exactly should be
sent in the state field (they say you can send anything) if no state is
part of a billing address so this commit mimics the behaviour of the
virtual terminal which uses the text "Outside of US".

Moreover Sage now supports countries outside of the US so it would
appear as though the supported_countries field could be updated.

Closes #2340