GlobalCollect: Set REJECTED refunds as unsuccessful transactions
It should be noted that Global Collect requires 24 hours for purchase to
settle before attempting a refund so remote tests for refunds are not
easily reproducible. Furthermore, forcing a REJECTED response in a
refund is not covered in the docs which is why this is lacking a remote
tests. However the unit test does take into account the response as shown
in the docs.

Closes #2365