JetPay: Pass down authorization payment method token to refund a capture
Capture responses do not return the token used during the authorize and
without a token, it's not possible to refund the capture without sending
in the original credit card number. This passes down the payment method
token returned in the authorization response to the capture's concatenated
authorization. This allows the token to be passed when refunding a
capture which negates the need to send in a credit card number.

Previously, this discrepency bubbled up as a failed refund with the
error message being that the credit card number is missing.