Revert "JetPay: Update XML API to 2.2"
This reverts commit 4a13ef6c3750853b3e589a545c4fc9377261471f.

Unfortunately upgrading to Jetpay's v2.2 API requires a certification
process that is not apparent while testing transactions against the
sandbox. Once certification happens, merchants are given a couple extra
fields that act as additional authentication values for requests. The
sandbox does not/did not surface those requirements.

This is being reverted for now to prevent any users of the Jetpay
adapter from confusion or accidental upgrades. We're currently speaking
with Jetpay support and reassessing the v2.2 upgrade to find the best
path forward. At this point it seems likely that we will need to break
off to a new JetpayV2 adapter similar to how the Checkout.com version 2
API was supported with a CheckoutV2 adapter.

It should be noted that the original intent to upgrade was to allow
merchants to pass in level 2 data for processing which is not available
in the current Jetpay API version in Active Merchant.